More deletion/moving of declarations sprinkled everywhere. Reindented p_weapon.c.

+++ source/p_view.c
@@ -362,7 +362,7 @@ void SV_CalcViewOffset (edict_t * ent)
 	ratio = (ent->client->fall_time - level.time) / FALL_TIME;
 	if (ratio < 0)
 		ratio = 0;
-	v[2] -= ratio * ent->client->fall_value * 0.4;
+	v[2] -= ratio * ent->client->fall_value * 0.4f;
 
 	// add bob height
 	bob = bobfracsin * xyspeed * bob_up->value;
@@ -1352,7 +1352,7 @@ void ClientEndServerFrame (edict_t * ent)
 	//
 	//if (FRAMESYNC)
 	{
-		xyspeed = sqrt(ent->velocity[0]*ent->velocity[0] + ent->velocity[1]*ent->velocity[1]);
+		xyspeed = sqrtf(ent->velocity[0]*ent->velocity[0] + ent->velocity[1]*ent->velocity[1]);
 
 		if (xyspeed < 5 || ent->solid == SOLID_NOT)
 		{
@@ -1378,7 +1378,7 @@ void ClientEndServerFrame (edict_t * ent)
 		current_client->bobtime += bobmove;
 
 		bobcycle = (int) current_client->bobtime;
-		bobfracsin = fabs (sin (current_client->bobtime * M_PI));
+		bobfracsin = fabsf (sinf (current_client->bobtime * M_PI));
 	}
 
 	// detect hitting the floor
